+++ CHANGELOG.md
@@ -64,4 +64,6 @@
 - fix casting citext column type (fix #2818) (#3861)
 - Add downgrade command (close #1156) (#3760)
 - persist mix files only when coverage is enabled (#3844)
+- `HASURA_GRAPHQL_EVENTS_FETCH_INTERVAL` changes semantics slightly: we only sleep for the interval
+  when there were previously no events to process. Potential space leak fixed. (#3839)
 - fix postgres query error when computed fields included in mutation response (fix #4035)